+++ src/build_number.h
@@ -1,1 +1,1 @@
-#define BUILD 1275
+#define BUILD 1276